Add failing specs for Dredd::Rack::Runner command generator
Please do notice that the Object#method has been removed from
the Dredd::Rack::Runner object in order to make place for the Dredd
`--method` option wrapper. That's a deliberate choice considering that
the Object#method is unlikely to be as useful as the option wrapper
in the Dredd::Rack::Runner context.

See http://ruby-doc.org/core-2.2.0/Object.html#method-i-method