Add support for Dredd::Rack::Runner automation
The Dredd::Rack::Runner can now be used to perform local
API blueprint validation programatically.

Any Rack application can be tested without having to start
any server manually. When performing a local validation,
the runner will start a Capybara::Server instance to serve
the application.

That allows to use the API blueprint as an acceptance test suite
to pratice BDD while the developers of the API clients use Apiary
as a mock server.