Remove Gemfile.lock
The Gemfile.lock in a gem is only used during development, it is
arguably not essential to lock dependencies in development, and
having the file causes automatic security scanners to trigger
most often that not alerts that are not relevant.

I used to keep the file in place in order to get a heads up on
current security alerts in general, I guess it would be better
if I was doing so independently of this code when the alerts do not
actually affect it.